python_requirements.txt
Django==1.10.3
django-appconf==1.0.2
django-compressor==2.1
django-pipeline==1.6.9
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
drf-nested-routers==0.11.1
PyJWT==1.4.2
rcssmin==1.0.6
rjsmin==1.0.12
